Added Maps API Key Variable
There are now 2 methods to load a Maps API key into the application. One way is to put it into an environment variable called MapsAPIKey. The other is to put it into the app.config file in the root of the application directory. I still need to add a prompt to appear when the API key isn't present.