Selecting and Loading Saved Plans
This was a sizable commit. I added Ikonli so I can render some icons easily. Added a saved plan fxml component to the start page for each saved plan. Clicking the plan will open the plan editor for that plan.

Next up is the waypoints to add. That is going to be sizable as well but I hope that the lessons I learned with this commit makes the next one easier.