Added Java Client Library for Maps
I didn't plan on using this library originally because it is a bit outdated when it comes to the new APIs like Routes. But it is still helpful in accomplishing a particular task of getting directions and converting the encoded polyline into a list of waypoints. I will be using those waypoints to search for Places along the route to add to the overall itinerary.