Simplified Plan Editor
This will be my final commit to this for a bit. I want to start working on other projects. I was going to add a method to find locations to visit along the path to the destination but I decided to put it on hold as it would quiet a bit more work to get right.

This commit changes some of the functions to rely on the PlanEditorSimplified.fxml and Controller so the app doesn't appear to be missing any features.